refactor tomopy shepp3d funcs to imcommon
